Create Saleshub API
This script sends a GET request to the Saleshub API with a specified set of headers and parameters. The headers include an authorization token and the content type, and the parameters specify a start and end date for the data to be returned. If the request is successful (status code 200), the script parses the response data as JSON and prints it to the console. If the request is not successful, it prints the status code to the console.